A load see description:
Changed the sudoku to show black boxes in staid of zeros.

Added an acknowledgement popup to ensure the right input format.

Changed the alert that the sudoku is not solvable form console output,
to UI.

Fixed the sudoku grid’s sizing.

Added a page to show the research.

Added a page to show our team.

Removed the old React.js logo.

Some language fixes.